Reorganize `snark-relations` module
